Refactor SearchFormMixin
